requirements.txt
llvmlite==0.23.0
numba==0.38.0
numpy==1.14.3
Pillow==5.1.0
tqdm==4.23.4
